requirements.txt
# Core dependencies for RAG system
requests>=2.28.0
beautifulsoup4>=4.11.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0

# Text processing
pdfplumber>=0.7.0
lxml>=4.9.0

# LLM support
openai>=1.0.0
transformers>=4.21.0
torch>=1.12.0

# Utilities
python-dotenv>=0.19.0
tqdm>=4.64.0
scikit-learn>=1.1.0

# Optional: Better display (for demo)
rich>=12.0.0
